+++ packages/block-editor/src/layouts/constrained.js
@@ -75,6 +75,8 @@ export default {
 						<div className="block-editor-hooks__layout-controls">
 							<div className="block-editor-hooks__layout-controls-unit">
 								<UnitControl
+									// TODO: Switch to `true` (40px size) if possible (https://github.com/WordPress/gutenberg/pull/64520#discussion_r1717314262)
+									__next40pxDefaultSize={ false }
 									className="block-editor-hooks__layout-controls-unit-input"
 									label={ __( 'Content' ) }
 									labelPosition="top"
@@ -96,6 +98,8 @@ export default {
 							</div>
 							<div className="block-editor-hooks__layout-controls-unit">
 								<UnitControl
+									// TODO: Switch to `true` (40px size) if possible
+									__next40pxDefaultSize={ false }
 									className="block-editor-hooks__layout-controls-unit-input"
 									label={ __( 'Wide' ) }
 									labelPosition="top"
@@ -125,6 +129,7 @@ export default {
 				) }
 				{ allowJustification && (
 					<ToggleGroupControl
+						__next40pxDefaultSize
 						__nextHasNoMarginBottom
 						label={ __( 'Justification' ) }
 						value={ justifyContent }